Move shared selector behind contract namespace
