+++ src/contract/mod.rs
@@ -1,0 +1,6 @@
+mod selection;
+
+pub(crate) use selection::{
+    SelectionError, collect_effective_checks, collect_effective_fixers, resolve_named_checks,
+    resolve_named_fixers, resolve_profiles,
+};
