+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod check;
 pub mod config;
+mod contract;
 mod exec;
 pub mod fix;
 mod git;
-mod selection;
